requirements.txt
fastapi
uvicorn
opencv-python
tesseract
transformers
torch
numpy
pillow
python-multipart